requirements.txt
pyyaml
beautifulsoup4
html5lib
lxml
nose
numpy
opencv-python
six
pillow-simd
psutil
pycryptodome
pylzma
pyopenssl
requests
send2trash
service-identity
lz4
pysocks
matplotlib
twisted
